Create Events.js
Events page created. Here users can register for upcoming events. Upon successful registration, an ICS file will be downloaded for the user to add the event to their Calender. They will also receive an email confirmation for the event. A few past events from the gallery page will be displayed on the events page for reference, and the user can view more of the past events by clicking the "view more" button. A new database "eventsRegistration" was created on Firebase Firestore to handle and store all event registration.